requirements.txt
streamlit>=1.30.0
streamlit-drawable-canvas>=0.9.3
dashscope>=1.14.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
moviepy>=1.0.3
tenacity>=8.2.3
streamlit-lottie>=0.0.5
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
scipy>=1.11.0
cryptography>=41.0.0
